Added landing page and work/about sections to portfolio website
In this commit, I added a landing page with links to the work and about sections, as well as the work and about sections themselves. The work section includes two sample projects, The about section includes a profile picture and some basic information about me, which can be customized further.